better to multiplex ESes than counts
